fixed issue with recomputing mindex for same object when dimension of X changes
Note that this does not effect the cross validation in sklearn since sklearn creates clones for each object in cross validation, essentially recomputing the multiindex for each. This makes it inefficient for this class but there really isn't another way around.

Also added tests to test the GridSearchCV to make sure it is finding the right order. The only way it can do that is if is recomputes the multiindex correctly.